Another fix for settings type ambiguity

+++ VRCOSC.Game/Graphics/Containers/Screens/ModuleEditScreen/ModuleEditSettingsContainer.cs
@@ -61,7 +61,7 @@ public class ModuleEditSettingsContainer : FillFlowContainer
                     });
                     break;
 
-                case int:
+                case long:
                     settingsFlow.Add(new ModuleSettingIntContainer
                     {
                         Key = key,
